Prevent creating two boards. Calc heuristic once.
* Previously, a board was instantiated in State but overwritten
immediately. The initial instantiation had been removed.
* Now, the heuristic is calculated once in the constuctor and saved
instead of being called repeatedly. This means that State.endState MUST
be set before instantiating the first State, or else the heuristic
values will be wrong. There is currently no check for this, becuase I
coudln't find a clean way to implement it in the constructor.